requirements.txt
dj-database-url==0.5.0
Django==2.1.1
djangorestframework==3.10.2
psycopg2==2.8.3
pytz==2019.1
